Compute local traversability maps
The hazard detector now computes traversability maps
where the rover is at the center of said map and every
point is marked traversable, unless it corresponds to
a detected hazard.

The map is oriented in the rover frame.

Because the hazard detector has no 3D information,
the user has to adjust new configuration parameters
describing the location of the region of interest
with respect to the rover frame (e.g., distance to the
right side of the ROI is +1m, -1 to the left, ...).

In the map, 0 corresponds to traversable, 1 means hazard.